# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/discount-codes": {
      "post": {
        "operationId": "DiscountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscountCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      },
      "get": {
        "operationId": "DiscountsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      }
    },
    "/api/v1/discount-codes/{code}/validate": {
      "get": {
        "operationId": "DiscountsController_validate",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      }
    },
    "/api/v1/discount-codes/{id}": {
      "get": {
        "operationId": "DiscountsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      },
      "patch": {
        "operationId": "DiscountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiscountCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      },
      "delete": {
        "operationId": "DiscountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Discounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMS Homestay API",
    "description": "API multi-tenant cho PMS homestay/khách sạn nhỏ VN. Auth: `Authorization: Bearer <access_token>` + header `X-Tenant-Slug`. Sinh tĩnh bằng `pnpm --filter @pms/api openapi:export`.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDiscountCodeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDiscountCodeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}